requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
black==26.1.0
celery==5.3.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.2.0
django-filter==23.5
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-stubs==3.16.6
dotenv==0.9.9
drf-yasg==1.21.7
flake8==7.3.0
idna==3.11
inflection==0.5.1
isort==7.0.0
kombu==5.6.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
